+++ vitest.config.ts
@@ -3,7 +3,13 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     coverage: {
-      exclude: ["dist/**", "test/**", "test-d/**"],
+      exclude: [
+        "dist/**",
+        "src/cli.ts",
+        "src/codegen.ts",
+        "test/**",
+        "test-d/**",
+      ],
       include: ["src/**/*.ts"],
       provider: "v8",
       reporter: ["text", "json-summary"],